requirements.txt
cv2
numpy
matplotlib
tensorflow
pyttsx3
smtplib
MIMEText
MIMEMultipart
MIMEImage
os
mysql.connector
datetime
Picamera2
sleep
tomato
apple
Send_email
gTTS
playsound 